Change validation to use data type annotations
This requires more work to split out the local aliases to stdlib, encode
the missing validations, and add depreciations to the current hodge-podge,
to tighten down what is allowed. Many of the unchecked edge-cases will
result in a broken ntp config anyways, but just fixing it would elevate
issues in this area from broken ntp service to broken catalog.